Change RestTemplate to be a bean
