+++ src/main/java/net/smartcosmos/cluster/gateway/config/GatewayConfiguration.java
@@ -1,5 +1,10 @@
 package net.smartcosmos.cluster.gateway.config;
 
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.Collections;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.embedded.undertow.UndertowDeploymentInfoCustomizer;
 import org.springframework.boot.context.embedded.undertow.UndertowEmbeddedServletContainerFactory;
@@ -9,10 +14,18 @@ import org.springframework.cloud.netflix.ribbon.SpringClientFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpRequest;
+import org.springframework.http.client.ClientHttpRequestExecution;
+import org.springframework.http.client.ClientHttpRequestInterceptor;
+import org.springframework.http.client.ClientHttpResponse;
+import org.springframework.http.client.InterceptingClientHttpRequestFactory;
 import org.springframework.security.config.annotation.authentication.configuration.EnableGlobalAuthentication;
 import org.springframework.security.config.annotation.authentication.configurers.GlobalAuthenticationConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.util.Base64Utils;
+import org.springframework.web.client.RestTemplate;
 
 /**
  * Configuration class for Gateway.
@@ -39,7 +52,7 @@ public class GatewayConfiguration extends GlobalAuthenticationConfigurerAdapter 
     }
 
     /**
-     * Need to configure Undertow to allo non standard wrappers or the exception gets lost in the noice.  This way the
+     * Need to configure Undertow to allow non standard wrappers or the exception gets lost in the noice.  This way the
      * exception will bubble out to the custom error controller.
      *
      * @return a modified Undertow config factory
@@ -50,5 +63,41 @@ public class GatewayConfiguration extends GlobalAuthenticationConfigurerAdapter 
         UndertowEmbeddedServletContainerFactory factory = new UndertowEmbeddedServletContainerFactory();
         factory.addDeploymentInfoCustomizers((UndertowDeploymentInfoCustomizer) deploymentInfo -> deploymentInfo.setAllowNonStandardWrappers(true));
         return factory;
+    }
+
+    @Bean
+    @Autowired
+    public RestTemplate authServerRestTemplate(
+        RibbonClientHttpRequestFactory ribbonClientHttpRequestFactory,
+        AuthenticationServerConnectionProperties authServerConnectionProperties) {
+
+        List<ClientHttpRequestInterceptor> interceptors = Collections.<ClientHttpRequestInterceptor>singletonList(
+            new BasicAuthorizationInterceptor(authServerConnectionProperties.getName(),
+                                              authServerConnectionProperties.getPassword()));
+        return new RestTemplate(new InterceptingClientHttpRequestFactory(ribbonClientHttpRequestFactory, interceptors));
+    }
+
+    private static class BasicAuthorizationInterceptor implements ClientHttpRequestInterceptor {
+
+        private static final String BASIC_AUTHENTICATION_HEADER = "Basic ";
+        private final String username;
+        private final String password;
+
+        BasicAuthorizationInterceptor(String username, String password) {
+
+            this.username = username;
+            this.password = (password == null ? "" : password);
+        }
+
+        @Override
+        public ClientHttpResponse intercept(
+            HttpRequest request, byte[] body,
+            ClientHttpRequestExecution execution) throws IOException {
+
+            String token = Base64Utils.encodeToString((this.username + ":" + this.password).getBytes(StandardCharsets.UTF_8));
+            request.getHeaders()
+                .add(HttpHeaders.AUTHORIZATION, BASIC_AUTHENTICATION_HEADER + token);
+            return execution.execute(request, body);
+        }
     }
 }
